requirements.txt
Django==1.7.1
Pillow==2.6.1
argparse==1.2.1
django-endless-pagination==2.0
gunicorn==19.1.1
psycopg2==2.5.4
wsgiref==0.1.2
